🚧 Add more institution logos to the footer and create insitutions file
