+++ src/components/Footer/Footer.tsx
@@ -1,27 +1,5 @@
 import * as Styles from './styles';
-import logoUfrgsNegative from '@assets/logoUfgNegativo.png';
-import logoUfgNegative from '@assets/logoUfrgsNegativo.png';
-import logoSabanciNormal from '@assets/logoSabanciNormal.png'; //change
-
-interface Institution {
-	name: string;
-	logoImage: string;
-}
-
-const institutionLogoImages: Institution[] = [
-	{
-		name: 'Federal University of Rio Grande do Sul',
-		logoImage: logoUfrgsNegative,
-	},
-	{
-		name: 'Federal University of Goiás',
-		logoImage: logoUfgNegative,
-	},
-	{
-		name: 'Sabanci University',
-		logoImage: logoSabanciNormal,
-	},
-];
+import institutionLogoImages from './institutions';
 
 function Footer() {
 	return (
